refactor(sync): restructure add sync device page with MVC architecture
- Split monolithic page into separate components, controller, and model
- Extract UI components: AlternativeMethodsFooter, DiscoveredDeviceCard, ScanningStatusCard
- Create AddSyncDeviceController for business logic separation
- Move DiscoveredDevice model to dedicated file
- Update import paths in sync devices page and app routes
- Improve code organization and maintainability
